created separate advanced results middleware to handle all get request queries of any route
