+++ server/controllers/companies.js
@@ -8,80 +8,7 @@ const path = require('path');
 //@route GET /api/v1/companies
 //@access public
 exports.getCompanies = asyncHandler(async (req, res, next) => {
-	//filtering by query string
-	let query;
-
-	// Copy req.query
-	const reqQuery = { ...req.query };
-
-	// Fields to exclude from query string
-	const removeFields = ['select', 'sort', 'page', 'limit'];
-
-	// Loop over removeFields and delete them from reqQuery
-	removeFields.forEach((param) => delete reqQuery[param]);
-
-	// Create query string
-	let queryStr = JSON.stringify(reqQuery);
-
-	// Create operators ($gt, $gte, etc)
-	queryStr = queryStr.replace(
-		/\b(gt|gte|lt|lte|in)\b/g,
-		(match) => `$${match}`
-	);
-
-	// Finding resource
-	query = Company.find(JSON.parse(queryStr));
-
-	//select fields
-	if (req.query.select) {
-		// the fields given with commas are seperated with spaces
-		const fields = req.query.select.split(',').join(' ');
-		query = query.select(fields);
-	}
-
-	// Sort
-	if (req.query.sort) {
-		const sortBy = req.query.sort.split(',').join(' ');
-		query = query.sort(sortBy);
-	} else {
-		query = query.sort('-createdAt');
-	}
-
-	// Pagination
-	const page = parseInt(req.query.page, 10) || 1;
-	const limit = parseInt(req.query.limit, 10) || 1;
-	const startIndex = (page - 1) * limit;
-	const endIndex = page * limit;
-	const total = await Company.countDocuments();
-
-	query = query.skip(startIndex).limit(limit);
-
-	// Executing query
-	const companies = await query;
-
-	// Pagination result
-	const pagination = {};
-
-	if (endIndex < total) {
-		pagination.next = {
-			page: page + 1,
-			limit,
-		};
-	}
-
-	if (startIndex > 0) {
-		pagination.prev = {
-			page: page - 1,
-			limit,
-		};
-	}
-
-	res.status(200).json({
-		success: true,
-		count: companies.length,
-		pagination: pagination,
-		data: companies,
-	});
+	res.status(200).json(res.advancedResults);
 });
 
 //@desc Get a single company
